Add deduplication feature
When a name conflict occurs, we can now check whether the files
are duplicates. This is done by checking file size to avoid
reading file contents for files which are obviously different
due to differing by size. Failing that, we compare MD5 hashes.
If the files match, then the source is simply removed instead
of adding another copy of the file to the destination.

Deduplication is enabled by default in mvr; use --no-dedupliate
to disable this behaviour.

The mvr subroutine now takes a deduplicate argument to control
whether this behaviour is enabled (by default, it is not).