Require Path::Tiny 0.034 and use its new features
Path::Tiny 0.027 added the digest method to produce a hex SHA-256
digest of the file. We can use that to make life easier.

Path::Tiny 0.034 replaced autodie's exceptions with Path::Tiny::Error.
This lets Path::Tiny remove the autodie exception, but now we have
to either handle the earlier autodie exceptions and the new
Path::Tiny::Error exceptions, or simply bump the version required.
I chose the latter, as the elimination of autodie from our prereq
tree is a nice-to-have.